Support for multiple pages of graphs.
On my display I can only fit about four graphs before they start to get too
small to read.  This commit introduces a new command-line switch:

  --graphs-per-page N

which specifies the maximum number of graphs to show on a single page.  After
this number is exceeded, a new tabbed-pane is created to hold the next N
graphs, and so on.

Also added an "auto-cycle" option, which cycles between tabs at a specifiable
duration.  I'm hoping this will make it easy to monitor lots of graphs without
having to juggle multiple windows and the like.